# Convert token parser to an async generator function