--- a/src/metadata-parser.ts
+++ b/src/metadata-parser.ts
@@ -127,7 +127,7 @@ function metadataParse(parser: Parser, options: InternalConnectionOptions, callb
         const type: DataType = TYPE[typeNumber];
 
         if (!type) {
-          return parser.emit('error', new Error(sprintf('Unrecognised data type 0x%02X', typeNumber)));
+          throw new Error(sprintf('Unrecognised data type 0x%02X', typeNumber));
         }
 
         switch (type.name) {
@@ -328,7 +328,7 @@ function metadataParse(parser: Parser, options: InternalConnectionOptions, callb
             });
 
           default:
-            return parser.emit('error', new Error(sprintf('Unrecognised type %s', type.name)));
+            throw new Error(sprintf('Unrecognised type %s', type.name));
         }
       });
     });