This pull request converts the token parsing code from a `Transform` stream into a async generator function that consumes a `AsyncIterable<Buffer>` and yields back parsed tokens.

All of the existing token parsing implementation is retained, this mostly just switches out how the parsing is "driven". This gives us the option to call other async code during the parsing (e.g. as is required for always encrypted to get encryption keys).

Right now, the async generator is wrapped in a `Readable.from` call, so we can keep most of the existing code as-is.

This depends on https://github.com/tediousjs/tedious/pull/1226 to be merged first.

---

### Performance

This change has a negative impact on performance, mostly because we introduce some additional overhead by having a async generator and wrapping it back into a `Readable` stream via `Readable.from`. This is an interim solution and will hopefully be refactored away soon, which should improve performance again. In the mean time, I believe the performance hit is acceptable.

#### Before

```
root ➜ ~/workspace/tedious (master) $ node benchmarks/query/select-many-rows.js 
query/select-many-rows.js size=10 n=10: 532.4236697461165 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=10: 284.16636656330303 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=10: 85.51023654585572 (minor: 3 - 3.2305240000000004ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=10: 15.934764679625893 (minor: 24 - 17.215927ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=100: 674.8609042503098 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=100: 463.037634212558 (minor: 3 - 3.7732279999999996ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=100: 147.4007835928836 (minor: 27 - 15.749910000000002ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=100: 20.14078649029258 (minor: 235 - 112.42549800000005ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=1000: 895.6707557867154 (minor: 6 - 14.221501ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=1000: 635.4394020000775 (minor: 31 - 30.979024999999993ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=1000: 166.43689997146637 (minor: 267 - 136.60967799999997ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=1000: 20.153075716417455 (minor: 2371 - 1165.8097820000007ms, major: 0 - 0ms, incremental: 0 - 0ms)
```

#### After

```
root ➜ ~/workspace/tedious (arthur/token-parser-async-generator) $ node benchmarks/query/select-many-rows.js 
query/select-many-rows.js size=10 n=10: 502.91585584057805 (minor: 0 - 0ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=10: 227.16272888405445 (minor: 1 - 2.704519ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=10: 81.94501407610478 (minor: 3 - 3.8326270000000005ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=10: 13.575519515447352 (minor: 32 - 22.000658ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=100: 524.0109713434697 (minor: 1 - 1.42641ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=100: 404.50896107083423 (minor: 5 - 8.497362ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=100: 107.47463633511748 (minor: 34 - 35.75296500000001ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=100: 16.49632461244278 (minor: 320 - 232.55980199999993ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10 n=1000: 854.1123887852357 (minor: 9 - 18.826438ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=100 n=1000: 539.9230060830604 (minor: 43 - 57.258018ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=1000 n=1000: 143.33281997154975 (minor: 344 - 222.09660600000004ms, major: 0 - 0ms, incremental: 0 - 0ms)
query/select-many-rows.js size=10000 n=1000: 15.99416653285739 (minor: 3203 - 2346.3768180000034ms, major: 0 - 0ms, incremental: 0 - 0ms)
```