@arthurschreiber 2021-02-26
> Some tests that used to call `parser.write` needs to be updated. I'll try to update those.

Ah, yes, you're right!

To update these tests, I think you should update the `parseTokens` function to also accept an `Iterable<Buffer>`, like this:

```diff
diff --git a/src/token/stream-parser.ts b/src/token/stream-parser.ts
index 85f27e3..5da37ad 100644
--- a/src/token/stream-parser.ts
+++ b/src/token/stream-parser.ts
@@ -47,25 +47,12 @@ class Parser {
   suspended: boolean;
   next?: () => void;
 
-  static async *parseTokens(iterable: AsyncIterable<Buffer>, debug: Debug, options: InternalConnectionOptions) {
-    const iterator = iterable[Symbol.asyncIterator]();
-
+  static async *parseTokens(iterable: AsyncIterable<Buffer> | Iterable<Buffer>, debug: Debug, options: InternalConnectionOptions) {
     const parser = new Parser(debug, options);
 
     let token: Token | undefined;
 
-    while (true) {
-      const result = await iterator.next();
-      if (result.done) {
-        if (parser.position === parser.buffer.length) {
-          return;
-        }
-
-        throw new Error('unexpected end of data');
-      }
-
-      const input = result.value;
-
+    for await (const input of iterable) {
       if (parser.position === parser.buffer.length) {
         parser.buffer = input;
       } else {
@@ -113,6 +100,10 @@ class Parser {
         }
       }
     }
+
+    if (parser.suspended) {
+      throw new Error('unexpected end of data');
+    }
   }
 
   constructor(debug: Debug, options: InternalConnectionOptions) {
```

This change will allow you to pass in an array of buffers to `parseTokens` in the test code (arrays implement the `Iterable` interface). The test cases will need to be converted to async functions, which will allow you to just pull the parsed tokens out of the async generator. Something like this:

```diff
diff --git a/test/unit/token/colmetadata-token-parser-test.js b/test/unit/token/colmetadata-token-parser-test.js
index 64416a9..b441d22 100644
--- a/test/unit/token/colmetadata-token-parser-test.js
+++ b/test/unit/token/colmetadata-token-parser-test.js
@@ -4,7 +4,7 @@ const TokenStreamParser = require('../../../src/token/stream-parser');
 const assert = require('chai').assert;
 
 describe('Colmetadata Token Parser', () => {
-  it('should int', () => {
+  it('should int', async function() {
     const numberOfColumns = 1;
     const userType = 2;
     const flags = 3;
@@ -20,11 +20,15 @@ describe('Colmetadata Token Parser', () => {
     buffer.writeBVarchar(columnName);
     // console.log(buffer.data)
 
-    const parser = new TokenStreamParser({ token() { } }, {}, {});
-    parser.write(buffer.data);
-    const token = parser.read();
-    // console.log(token)
+    const parser = TokenStreamParser.parseTokens([buffer.data], { token() { } }, {}, {});
+    const tokens = [];
+    for await (const token of parser) {
+      tokens.push(token);
+    }
+
+    assert.lengthOf(tokens, 1);
 
+    const token = tokens[0];
     assert.isOk(!token.error);
     assert.strictEqual(token.columns.length, 1);
     assert.strictEqual(token.columns[0].userType, 2);
```