@arthurschreiber 2021-02-26
> To update these tests, I think you should update the `parseTokens` function to also accept an `Iterable<Buffer>`, like this:

@IanChokS I pushed the change I described here - but I didn't update any of the tests as you mentioned you want to look into that.