@IanChokS 2021-02-26
Oh I already started reworking the tests with the previous implementation of the parser, which looks like this:

```js
  it('should be of bad type', (done) => {
    const buffer = new WritableTrackingBuffer(50, 'ucs2');

    buffer.writeUInt8(0xe3);
    buffer.writeUInt16LE(0); // Length written later
    buffer.writeUInt8(0xff); // Bad type

    const data = buffer.data;
    data.writeUInt16LE(data.length - 3, 1);

    const asyncIterable = {
      async*[Symbol.asyncIterator]() {
        yield buffer.data;
      }
    };

    const parser = StreamParser.parseTokens(asyncIterable, {}, {});

    (async () => {
      for await (let token of parser) { // (4)
        assert.strictEqual(token, null);
        done();
      }
    })()
  });
```'
on top of that, I've added some slight changes to the old parser, one which includes adding an optional parser object instead of forcefully creating a new parser object in the `*parserTokens` method. This is to fix the `row-parser-tests` because the tests manually set `colMetadata` property in the `Parser` class which previously was only created inside the `*parseTokens` method and therefore was not possible to set the field. With this optional parameter, the tests now look something like this: 
```js
// Parameter change
static async *parseTokens(iterable: AsyncIterable<Buffer>, debug: StreamParser, options: InternalConnectionOptions, parser?: Parser) 
```

```js
// Updated row-token-parser test e.g.,
    const asyncIterable = {
      async*[Symbol.asyncIterator]() {
        yield buffer.data;
      }
    };

    const parserObj = new Parser({}, options);
    parserObj.colMetadata = colMetadata;
    const parser = Parser.parseTokens(asyncIterable, {}, options, parserObj);

    (async () => {
      for await (let token of parser) {
        assert.strictEqual(token.columns.length, 1);
        assert.strictEqual(token.columns[0].value, value);
        assert.strictEqual(token.columns[0].metadata, colMetadata[0]);
        done();
      }
    })()
```
Are these changes acceptable so far?