@IanChokS 2021-02-26
Actually with the option `Iterable<Buffer>`, it would be easier to just pass in the Buffer arrays for the tests instead of creating an AsyncIterable object like 
```js
    const asyncIterable = {
      async*[Symbol.asyncIterator]() {
        yield buffer.data;
      }
    };
```
I'll pull your changes and update the tests accordingly