@arthurschreiber 2021-02-26
@IanChokS that parameter change to allow injecting a pre-built parser object isn't great, but it's okay for now to get the tests fixed up! 👍 I hope that can be refactored away sooner or later anyway. 😄