@arthurschreiber 2021-02-26
> Actually with the option `Iterable<Buffer>`, it would be easier to just pass in the Buffer arrays for the tests instead of creating an AsyncIterable object like

Yes, that's the main idea. 😅

A simpler version of the approach you've taken to build an async iterable object is to define an async generator function and call it like this:

```ts
const asyncIterable = (async function*() => {
  yield buffer.data;
})();
```

But yes, for the tests I think just using an array of buffers is just simpler and easier to read and understand. 👍