@arthurschreiber 2021-02-26
I pushed one more change that _slightly_ improves performance by re-using the callback that's passed into the individual token parsing functions. The performance difference isn't super big, but it's an easy improvement.